Started socket infrastructure
