+++ LLS/LLS.py
@@ -3,6 +3,7 @@ import random
 import sys
 import os
 import typing
+from dsast import DSAST
 # import base64
 # import hashlib
 # from Crypto import Random
@@ -37,20 +38,19 @@ class GAST:
 		self.permissions = aOIT.permissions
 		self.domain = aOIT.domain
 		self.key = aOIT.key
+		aBAST = BAST() # address for tagged data
+		bastTable[(self.domain, self.key)] = aBAST
 
-		# TODO: add a function to add entry to OIT tree/table
 		# if(encrypt):
 		# 	# aOAT = OAT(aOIT)
 		# 	# gastTable[(self.domain, self.key)] = aOAT.value # not sure about this part, dave said to ignore OAT for the moment
-		# 	aBAST = BAST()
-		# 	bastTable[(self.domain, self.key)] = aBAST # BAST is the GAST's value
-		# 	gastTable[aBAST] = self
 		# else:
-		aBAST = BAST() # address for tagged data
-		bastTable[(self.domain, self.key)] = aBAST
+		# aBAST = BAST() # address for tagged data
+		# bastTable[(self.domain, self.key)] = aBAST
 		# TODO: also have to add a corresponding BAST/OAT
 		
-	# TODO: external calls can create, duplicate, and free a particular GAST
+# TODO: external calls can create, duplicate, and free a particular GAST
+#		Maybe should be LLS exclusive functions.
 	def duplicate(self):
 		newGAST = GAST(permissions=self.permissions)
 		bastTable[(newGAST.domain, newGAST.key)] = bastTable[(self.domain, self.key)]
@@ -151,28 +151,7 @@ class OIT:
 		# need to check if key does not exist in domain before addying entry to hash table
 		# not sure if this is what goes into the oitTable...
 		if (self.domain, self.value) not in oitTable:
-			oitTable.append((self.domain, self.value))
-
-class DSAST:
-	# Structure(MSB to LSB): Way Limit, Size, Line Limit, Index, Offset, prob useless for python simulations
-	def __init__(self, index=0, lineLimit=0, size=0, wayLimit=0):
-		# Size: 0 = Large DSAST, 1 = Small DSAST
-		# Offset: 40 bits for small, 50 bit for large
-		# Index: (mb kinda the file name) indexing in a list of DSASTs
-		# only cache fields
-		# Line limit: The least significant set-bit of the this field determines the partition size and the remaining 
-		# upper bits of the field determine to which of the possible ranges, of the indicated size, the DSAST is restricted.
-		# Way Limit: cache restrictions. 0=RESERVED, 1=no restriction, 2=only 1st half of the cache, 3=only second half of the cache
-
-		self.size = size
-		self.index = index
-		self.linelimit = lineLimit
-		self.waylimit = wayLimit
-		if(size): # small DSAST
-			self.offset = random.getrandbits(40)
-		else:
-			self.offset = random.getrandbits(50)
-		
+			oitTable.append((self.domain, self.value))		
 
 def mapBASTtoDSAST(dsast, gast=None):
 	global bastTable
