+++ lib/sockets.txt
@@ -1,0 +1,1 @@
+# put sockets for each component here
